+++ CHANGES.md
@@ -7,6 +7,7 @@ Next
 * Fixed the new module form, it failed if the user wasn't specified.
 * Fixed the show module page, it failed if the feed URL was invalid.
 * Fixed mailer to generate URLs that include port and protocol if necessary.
+* Fixed flash notification message helper to ensure correct arguments, added docs and specs.
 * Switched to 'maruku' gem, a pure Ruby Markdown parser.
 * Added `rake spec:rcov:save` to save code coverage data, and `rake spec:rcov:diff` to display uncovered code since save.
 * Added `cap db:use` to download remote `production` database and replace local `development` database with it.